Removed rapidjson from all public headers. Its now only used for reading the json and as the default writer. This allows us to use a different version or other json library if we specify a custom writer from our own app.
