+++ COLLADA2GLTF/GLTF/GLTFAccessor.cpp
@@ -26,7 +26,6 @@
 
 #include "GLTF.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
