+++ COLLADA2GLTF/GLTF/GLTFAnimation.cpp
@@ -26,8 +26,6 @@
 #include "animationConverter.h"
 #include "GLTF-Open3DGC.h" //FIXME: setupAndWriteAnimationParameter has to move out from this header
 
-using namespace rapidjson;
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
