+++ COLLADA2GLTF/GLTF/GLTFBuffer.cpp
@@ -27,8 +27,6 @@
 #include "GLTF.h"
 #include "GLTFAsset.h"
 
-using namespace rapidjson;
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
