+++ COLLADA2GLTF/GLTF/GLTFConfig.cpp
@@ -23,7 +23,6 @@
 
 #include "GLTF.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
