+++ COLLADA2GLTF/GLTF/GLTFMesh.cpp
@@ -28,7 +28,6 @@
 #include "GLTF.h"
 #include "../helpers/geometryHelpers.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
