+++ COLLADA2GLTF/assetModifiers/GLTFFlipUVModifier.cpp
@@ -24,7 +24,6 @@
 #include "GLTF.h"
 #include "GLTFFlipUVModifier.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
