+++ COLLADA2GLTF/convert/animationConverter.cpp
@@ -8,7 +8,6 @@
 
 #include "GLTF-Open3DGC.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
