+++ COLLADA2GLTF/helpers/encodingHelpers.cpp
@@ -24,7 +24,6 @@
 #include "GLTF.h"
 #include "encodingHelpers.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
