+++ COLLADA2GLTF/helpers/mathHelpers.cpp
@@ -25,7 +25,6 @@
 #include "../GLTFOpenCOLLADA.h"
 #include "mathHelpers.h"
 
-using namespace rapidjson;
 #if __cplusplus <= 199711L
 using namespace std::tr1;
 #endif
